requirements.txt
flask
python-dotenv
psycopg2-binary
requests
pydub
vosk
cyton
'nemo_toolkit[all]'